# api/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for TodoApp",
        "title": "Todo API",
        "termsOfService": "localhost:8080",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/tasks": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "検索ワードとstatusの指定でフィルターをかけてTodoの一覧を取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Task一覧を取得する",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "ページ数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "ページサイズ",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "Pending",
                        "description": "Todoのステータス(Pending, InProgress, Completed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Todoのタイトルまたは説明文から検索する検索ワード",
                        "name": "searchWord",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.getTaskResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "パラメーターから新規Todoを作成する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Taskを登録する",
                "parameters": [
                    {
                        "description": "Task登録",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.CreateTaskParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/task.createTaskResponse"
                        }
                    }
                }
            }
        },
        "/v1/tasks/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "指定されたIDに合致するTodoのタイトルやステータスを更新する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Taskを更新する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "更新するTodoを指定するid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task更新",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.UpdateTaskParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.updateTaskResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "指定されたIDに合致するTodoを削除する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Taskを削除する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "削除するTodoを指定するid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.deleteTaskResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "task.CreateTaskParams": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "「達人に学ぶクリーンアーキテクチャp100~105」までを読む"
                },
                "title": {
                    "type": "string",
                    "example": "読書"
                }
            }
        },
        "task.Task": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "task.UpdateTaskParams": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1,
                    "example": "「達人に学ぶクリーンアーキテクチャp200~300」までを読む"
                },
                "status": {
                    "type": "string",
                    "example": "Completed"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "輪読会"
                }
            }
        },
        "task.createTaskResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "4082ed31-263c-40ec-9d41-e9d274c6bca8"
                }
            }
        },
        "task.deleteTaskResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "4082ed31-263c-40ec-9d41-e9d274c6bca8"
                }
            }
        },
        "task.getTaskResponse": {
            "type": "object",
            "properties": {
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/task.Task"
                    }
                },
                "totalTasks": {
                    "type": "string"
                }
            }
        },
        "task.updateTaskResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "4082ed31-263c-40ec-9d41-e9d274c6bca8"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Todo-API-Key",
            "in": "header"
        }
    }
}